refactor: eliminate explorer_helpers.rs by distributing functions to proper layers
Move AppState-only functions to service/explorer_state.rs:
- get_tab_index_by_key
- get_active_tab_state_query
- get_tab_state_query_by_key
- Add get_tab_state_by_index (new)
- Add update_tab_state (new)

Move emit operations to app/explorer.rs as internal helpers:
- emit_tab_state (new)
- update_tab_and_emit (moved from helper module)
- emit_current_tab_state (moved from helper module)

This aligns with layer design:
- service/ = AppState operations only
- app/ = AppHandle/Emitter operations
