+++ src-tauri/src/service/explorer_types.rs
