+++ src-tauri/src/service/mod.rs
@@ -1,5 +1,6 @@
 pub mod app_state;
 pub mod explorer_state;
+pub mod explorer_types;
 pub mod types;
 pub mod viewer_state;
